requirements.txt
streamlit
openai
textblob
vaderSentiment
python-dotenv
